+++ src/main/java/com/example/services/BalanceService.java
@@ -3,5 +3,5 @@ package com.example.services;
 import com.example.models.Request;
 
 public interface BalanceService {
-    int getBalance(Request request);
+    double getBalance(Request request);
 }
